+++ carp_webservices/test/carp_service_test.dart
@@ -190,7 +190,8 @@ void main() {
     });
 
     test('- query', () async {
-      String query = 'carp_header.user_id==$userId';
+      String query = 'carp_header.user_id==$userId;carp_body.timestamp>2019-11-02T12:53:40.219598Z';
+      //String query = 'carp_header.user_id==$userId';
       print("query : $query");
       List<CARPDataPoint> data = await CarpService.instance.getDataPointReference().queryDataPoint(query);